pyproject.toml
[project]
name = "filegram"
version = "0.1.0"
description = "A Python code agent powered by multiple LLM providers with MCP support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
    "httpx>=0.25.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    "mcp>=1.0.0",
    "pymupdf>=1.24.0",  # PDF reading support
    # Benchmark dependencies
    "networkx>=3.0",     # zep adapter: knowledge graph
    "mem0ai>=1.0.0",     # mem0 adapter: memory framework
    "cohere>=5.0.0",     # naive_rag & mem0: Cohere embed-v4.0
    "numpy>=1.21",       # naive_rag: embedding similarity
    "azure-identity>=1.15.0",  # mem0: Azure OpenAI LLM provider
]

[project.scripts]
filegramengine = "filegramengine.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["filegramengine"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.ruff.format]
quote-style = "double"
